backend/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
pytest
httpx
sqlalchemy
psycopg2-binary
python-jose
pydantic-settings